machine_learning/torch_anomaly_detection/pyproject.toml
[tool.poetry]
name = "torch_anomaly_detection"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
matplotlib = "^3.2.2"
numpy = "^1.19.0"
pillow = "^7.2.0"
python = "^3.7"
pytorch-lightning = "^0.8.5"
torch = "^1.5.1"
torchvision = "^0.6.1"
torchsummary = "^1.5.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.8.3"
mypy = "^0.782"
setuptools = "^49.1.2"
tensorflow = "^2.2.0"
tensorboard = "^2.2.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
